move variant class to new variant file to resolve issues with circular imports
