+++ happler/tree/haplotypes.py
@@ -7,87 +7,7 @@ import numpy.typing as npt
 
 from .tree import Tree
 from ..data import Genotypes
-
-
-class VariantType:
-    """
-    A class denoting the type of variant
-
-    Attributes
-    ----------
-    type : str, optional
-        The type of variant (ex: SNP, STR, etc)
-
-        Defaults to a single nucleotide polymorphism
-
-    """
-
-    def __init__(self, variant_type: str = "snp"):
-        # TODO: add support for STRs, SVs, etc
-        supported_types = ["snp"]
-        # a VariantType has a dtype of str
-        self.dtype = np.dtype("S5")
-        # store the type if it is supported
-        variant_type = variant_type.lower()
-        if variant_type in supported_types:
-            self.type = variant_type
-        else:
-            raise ValueError("{}s are not yet supported.".format(variant_type.upper()))
-
-    def __repr__(self):
-        return self.type.upper()
-
-    def __eq__(self, other):
-        return self.type == other.type
-
-
-# We declare this class to be a dataclass to automatically define __init__ and a few
-# other methods. We use frozen=True to make it immutable.
-@dataclass(frozen=True)
-class Variant:
-    """
-    A variant within the genotypes matrix
-
-    Attributes
-    ----------
-    id : str
-        The variant's unique ID
-    idx : int
-        The index of the variant within the genotype data
-    pos : int
-        The chromosomal start position of the variant
-    """
-
-    idx: int
-    id: int
-    pos: int
-
-    @classmethod
-    def from_np(cls, np_mixed_arr_var: np.void, idx: int) -> Variant:
-        """
-        Convert a numpy mixed array variant record into a Variant
-
-        Parameters
-        ----------
-        np_mixed_arr_var : np.void
-            A numpy mixed array variant record with entries 'id' and 'pos'
-        idx : int
-            See :py:attr:`~.Variant.idx`
-
-        Returns
-        -------
-        Variant
-            The converted Variant
-        """
-        return cls(idx=idx, id=np_mixed_arr_var["id"], pos=np_mixed_arr_var["pos"])
-
-    @property
-    def ID(self):
-        return self.id
-
-    @property
-    def POS(self):
-        return self.pos
+from .variant import Variant
 
 
 # We declare this class to be a dataclass to automatically define __init__ and a few
@@ -214,7 +134,15 @@ class Haplotypes:
     """
 
     def __init__(self):
-        self.data = np.array()
+        self.data = np.array(
+            [],
+            dtype=[
+                ("hap", np.uint),
+                ("tree", np.uint),
+                ("variant", "U50"),
+                ("score", np.float64),
+            ],
+        )
 
     @classmethod
     def from_tree(cls, tree: Tree) -> Haplotypes:
@@ -233,23 +161,3 @@ class Haplotypes:
                 ("score", np.float64),
             ],
         )
-
-    def add_variant(self, variant: Variant, allele: int, hap_idxs: list[int] = None):
-        # do we already have any haplotypes?
-        variant_gts = self.genotypes.data[:, variant.idx, allele][:, np.newaxis]
-        if self.haplotypes:
-            if hap_idxs is None:
-                hap_idxs = list(range(len(self.haplotypes)))
-            for hap_idx in hap_idxs:
-                self.haplotypes[hap_idx] = self.haplotypes[hap_idx].append(
-                    variant, allele, variant_gts
-                )
-            self.data[:, hap_idxs] = np.logical_and(self.data[:, hap_idxs], variant_gts)
-        else:
-            self.haplotypes = [Haplotype.from_node(variant, allele, variant_gts)]
-            self.data = variant_gts
-
-    # WHAT DO WE NEED THIS CLASS TO DO?
-    # 1) create a haplotype matrix where each (SNP, allele) pair is added to an existing haplotype
-    # - at the root of the tree, this is just the same as add_variant()
-    # 2) compose a vector of haplotype dosages from the haplotype matrix
